Crate sp_consensus_pow

Primitives for Substrate Proof-of-Work (PoW) consensus.

Constants

POW_ENGINE_ID

The ConsensusEngineId of PoW.

Traits

DifficultyApi

API for those chains that put their difficulty adjustment algorithm directly onto runtime. Note that while putting difficulty adjustment algorithm to runtime is safe, putting the PoW algorithm on runtime is not.

TimestampApi

API necessary for timestamp-based difficulty adjustment algorithms.

TotalDifficulty

Define methods that total difficulty should implement.

Type Aliases

Seal

Type of seal.